backend/requirements.txt
flask==3.1.2
flask-cors==6.0.1
pymongo==4.15.4
google-generativeai==0.8.5
pydantic==2.12.5


